refactored date inputs to share functionality
