+++ src/dpr/components/_filters/filter-input/enum.ts
@@ -2,6 +2,7 @@
 export enum FilterType {
   radio = 'Radio',
   dateRange = 'DateRange',
+  granularDateRange = 'GranularDateRange',
   date = 'Date',
   select = 'Select',
   autocomplete = 'Autocomplete',
